Split up useTests/summarizeTests
