+++ src/utils/useTests.test.ts
@@ -52,83 +52,6 @@ const tests = [
   }),
 ];
 
-it("Calculates static information correctly", () => {
-  const { isInProgress, changeCount, brokenCount, browserResults, viewportResults } =
-    useTests(tests);
-
-  expect({
-    isInProgress,
-    changeCount,
-    brokenCount,
-    browserResults,
-    viewportResults,
-  }).toMatchInlineSnapshot(`
-    {
-      "brokenCount": 1,
-      "browserResults": [
-        {
-          "browser": {
-            "id": "CHROME",
-            "key": "CHROME",
-            "name": "Chrome",
-            "version": "<unknown>",
-          },
-          "result": "CAPTURE_ERROR",
-        },
-        {
-          "browser": {
-            "id": "SAFARI",
-            "key": "SAFARI",
-            "name": "Safari",
-            "version": "<unknown>",
-          },
-          "result": "CAPTURE_ERROR",
-        },
-      ],
-      "changeCount": 1,
-      "isInProgress": true,
-      "viewportResults": [
-        {
-          "result": "EQUAL",
-          "viewport": {
-            "id": "_1200",
-            "isDefault": true,
-            "name": "1200px",
-            "width": 1200,
-          },
-        },
-        {
-          "result": "CAPTURE_ERROR",
-          "viewport": {
-            "id": "_800",
-            "isDefault": false,
-            "name": "800px",
-            "width": 800,
-          },
-        },
-        {
-          "result": "CHANGED",
-          "viewport": {
-            "id": "_480",
-            "isDefault": false,
-            "name": "480px",
-            "width": 480,
-          },
-        },
-        {
-          "result": "ADDED",
-          "viewport": {
-            "id": "_1600",
-            "isDefault": false,
-            "name": "1600px",
-            "width": 1600,
-          },
-        },
-      ],
-    }
-  `);
-});
-
 it("Picks and updates selectedTest when choosing viewport", () => {
   const onSelectTest = jest.fn((x) => {});
   jest.mocked(useState).mockImplementationOnce(((x: any) => [x, onSelectTest]) as any);
